Extract nested DTO's to separate classes
