+++ bundles/org.openhab.binding.hdpowerview/src/main/java/org/openhab/binding/hdpowerview/internal/api/responses/Shade.java
@@ -14,7 +14,7 @@ package org.openhab.binding.hdpowerview.internal.api.responses;
 
 import org.eclipse.jdt.annotation.NonNullByDefault;
 import org.eclipse.jdt.annotation.Nullable;
-import org.openhab.binding.hdpowerview.internal.api.responses.Shades.ShadeData;
+import org.openhab.binding.hdpowerview.internal.api.ShadeData;
 
 /**
  * State of a single Shade, as returned by an HD PowerView hub
@@ -23,6 +23,5 @@ import org.openhab.binding.hdpowerview.internal.api.responses.Shades.ShadeData;
  */
 @NonNullByDefault
 public class Shade {
-
     public @Nullable ShadeData shade;
 }
