+++ bundles/org.openhab.binding.hdpowerview/src/test/java/org/openhab/binding/hdpowerview/AutomationChannelBuilderTest.java
@@ -24,10 +24,9 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.openhab.binding.hdpowerview.internal.HDPowerViewBindingConstants;
 import org.openhab.binding.hdpowerview.internal.HDPowerViewTranslationProvider;
-import org.openhab.binding.hdpowerview.internal.api.responses.SceneCollections.SceneCollection;
-import org.openhab.binding.hdpowerview.internal.api.responses.Scenes.Scene;
-import org.openhab.binding.hdpowerview.internal.api.responses.ScheduledEvents;
-import org.openhab.binding.hdpowerview.internal.api.responses.ScheduledEvents.ScheduledEvent;
+import org.openhab.binding.hdpowerview.internal.api.Scene;
+import org.openhab.binding.hdpowerview.internal.api.SceneCollection;
+import org.openhab.binding.hdpowerview.internal.api.ScheduledEvent;
 import org.openhab.binding.hdpowerview.internal.builders.AutomationChannelBuilder;
 import org.openhab.binding.hdpowerview.providers.MockedLocaleProvider;
 import org.openhab.binding.hdpowerview.providers.MockedTranslationProvider;
@@ -78,7 +77,7 @@ public class AutomationChannelBuilderTest {
 
     @Test
     public void sceneSunriseWeekends() {
-        ScheduledEvent scheduledEvent = createScheduledEventWithScene(ScheduledEvents.SCHEDULED_EVENT_TYPE_SUNRISE);
+        ScheduledEvent scheduledEvent = createScheduledEventWithScene(ScheduledEvent.SCHEDULED_EVENT_TYPE_SUNRISE);
         scheduledEvent.daySaturday = true;
         scheduledEvent.daySunday = true;
 
@@ -91,7 +90,7 @@ public class AutomationChannelBuilderTest {
 
     @Test
     public void sceneSunsetWeekdays() {
-        ScheduledEvent scheduledEvent = createScheduledEventWithScene(ScheduledEvents.SCHEDULED_EVENT_TYPE_SUNSET);
+        ScheduledEvent scheduledEvent = createScheduledEventWithScene(ScheduledEvent.SCHEDULED_EVENT_TYPE_SUNSET);
         scheduledEvent.dayMonday = true;
         scheduledEvent.dayTuesday = true;
         scheduledEvent.dayWednesday = true;
@@ -107,7 +106,7 @@ public class AutomationChannelBuilderTest {
 
     @Test
     public void sceneTimeAllDays() {
-        ScheduledEvent scheduledEvent = createScheduledEventWithScene(ScheduledEvents.SCHEDULED_EVENT_TYPE_TIME);
+        ScheduledEvent scheduledEvent = createScheduledEventWithScene(ScheduledEvent.SCHEDULED_EVENT_TYPE_TIME);
         scheduledEvent.dayMonday = true;
         scheduledEvent.dayTuesday = true;
         scheduledEvent.dayWednesday = true;
@@ -127,7 +126,7 @@ public class AutomationChannelBuilderTest {
 
     @Test
     public void sceneMinutesBeforeSunriseMondayTuesday() {
-        ScheduledEvent scheduledEvent = createScheduledEventWithScene(ScheduledEvents.SCHEDULED_EVENT_TYPE_SUNRISE);
+        ScheduledEvent scheduledEvent = createScheduledEventWithScene(ScheduledEvent.SCHEDULED_EVENT_TYPE_SUNRISE);
         scheduledEvent.dayMonday = true;
         scheduledEvent.dayTuesday = true;
         scheduledEvent.minute = -15;
@@ -141,7 +140,7 @@ public class AutomationChannelBuilderTest {
 
     @Test
     public void sceneHoursMinutesAfterSunriseMonday() {
-        ScheduledEvent scheduledEvent = createScheduledEventWithScene(ScheduledEvents.SCHEDULED_EVENT_TYPE_SUNRISE);
+        ScheduledEvent scheduledEvent = createScheduledEventWithScene(ScheduledEvent.SCHEDULED_EVENT_TYPE_SUNRISE);
         scheduledEvent.dayMonday = true;
         scheduledEvent.minute = 61;
 
@@ -154,7 +153,7 @@ public class AutomationChannelBuilderTest {
 
     @Test
     public void sceneMinutesBeforeSunsetWednesdayThursdayFriday() {
-        ScheduledEvent scheduledEvent = createScheduledEventWithScene(ScheduledEvents.SCHEDULED_EVENT_TYPE_SUNSET);
+        ScheduledEvent scheduledEvent = createScheduledEventWithScene(ScheduledEvent.SCHEDULED_EVENT_TYPE_SUNSET);
         scheduledEvent.dayWednesday = true;
         scheduledEvent.dayThursday = true;
         scheduledEvent.dayFriday = true;
@@ -169,7 +168,7 @@ public class AutomationChannelBuilderTest {
 
     @Test
     public void sceneHourAfterSunsetFridaySaturdaySunday() {
-        ScheduledEvent scheduledEvent = createScheduledEventWithScene(ScheduledEvents.SCHEDULED_EVENT_TYPE_SUNSET);
+        ScheduledEvent scheduledEvent = createScheduledEventWithScene(ScheduledEvent.SCHEDULED_EVENT_TYPE_SUNSET);
         scheduledEvent.dayFriday = true;
         scheduledEvent.daySaturday = true;
         scheduledEvent.daySunday = true;
@@ -185,7 +184,7 @@ public class AutomationChannelBuilderTest {
     @Test
     public void sceneCollection() {
         ScheduledEvent scheduledEvent = createScheduledEventWithSceneCollection(
-                ScheduledEvents.SCHEDULED_EVENT_TYPE_SUNRISE);
+                ScheduledEvent.SCHEDULED_EVENT_TYPE_SUNRISE);
 
         List<ScheduledEvent> scheduledEvents = new ArrayList<>(List.of(scheduledEvent));
         List<Channel> channels = builder.withSceneCollections(sceneCollections).withScheduledEvents(scheduledEvents)
@@ -197,7 +196,7 @@ public class AutomationChannelBuilderTest {
 
     @Test
     public void suppliedListIsUsed() {
-        ScheduledEvent scheduledEvent = createScheduledEventWithScene(ScheduledEvents.SCHEDULED_EVENT_TYPE_SUNRISE);
+        ScheduledEvent scheduledEvent = createScheduledEventWithScene(ScheduledEvent.SCHEDULED_EVENT_TYPE_SUNRISE);
         List<ScheduledEvent> scheduledEvents = new ArrayList<>(List.of(scheduledEvent));
         List<Channel> existingChannels = new ArrayList<>(0);
         List<Channel> channels = builder.withScenes(scenes).withScheduledEvents(scheduledEvents)
@@ -215,7 +214,7 @@ public class AutomationChannelBuilderTest {
 
     @Test
     public void emptyListWhenNoScenesOrSceneCollections() {
-        ScheduledEvent scheduledEvent = createScheduledEventWithScene(ScheduledEvents.SCHEDULED_EVENT_TYPE_SUNRISE);
+        ScheduledEvent scheduledEvent = createScheduledEventWithScene(ScheduledEvent.SCHEDULED_EVENT_TYPE_SUNRISE);
         List<ScheduledEvent> scheduledEvents = new ArrayList<>(List.of(scheduledEvent));
         List<Channel> channels = builder.withScheduledEvents(scheduledEvents).build();
 
@@ -225,7 +224,7 @@ public class AutomationChannelBuilderTest {
     @Test
     public void emptyListWhenNoSceneForScheduledEvent() {
         ScheduledEvent scheduledEvent = createScheduledEventWithSceneCollection(
-                ScheduledEvents.SCHEDULED_EVENT_TYPE_SUNRISE);
+                ScheduledEvent.SCHEDULED_EVENT_TYPE_SUNRISE);
         List<ScheduledEvent> scheduledEvents = new ArrayList<>(List.of(scheduledEvent));
         List<Channel> channels = builder.withScenes(scenes).withScheduledEvents(scheduledEvents).build();
 
@@ -234,7 +233,7 @@ public class AutomationChannelBuilderTest {
 
     @Test
     public void emptyListWhenNoSceneCollectionForScheduledEvent() {
-        ScheduledEvent scheduledEvent = createScheduledEventWithScene(ScheduledEvents.SCHEDULED_EVENT_TYPE_SUNRISE);
+        ScheduledEvent scheduledEvent = createScheduledEventWithScene(ScheduledEvent.SCHEDULED_EVENT_TYPE_SUNRISE);
         List<ScheduledEvent> scheduledEvents = new ArrayList<>(List.of(scheduledEvent));
 
         List<Channel> channels = builder.withSceneCollections(sceneCollections).withScheduledEvents(scheduledEvents)
@@ -245,7 +244,7 @@ public class AutomationChannelBuilderTest {
 
     @Test
     public void groupAndIdAreCorrect() {
-        ScheduledEvent scheduledEvent = createScheduledEventWithScene(ScheduledEvents.SCHEDULED_EVENT_TYPE_SUNRISE);
+        ScheduledEvent scheduledEvent = createScheduledEventWithScene(ScheduledEvent.SCHEDULED_EVENT_TYPE_SUNRISE);
         scheduledEvent.id = 42;
         List<ScheduledEvent> scheduledEvents = new ArrayList<>(List.of(scheduledEvent));
         List<Channel> channels = builder.withScenes(scenes).withScheduledEvents(scheduledEvents).build();
